fix(attention): address Gluon review feedback
